requirements.txt
pandas==1.3.3
numpy==1.19.5
pandas==1.3.3
logger
pillow
scikit-learn
tensorflow==2.5.0
requests==2.26.0
Flask==2.0.1
Flask-WTF==0.15.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
future==0.18.2
gunicorn==20.0.4
h2o==3.26.0.10
idna==2.9
itsdangerous==2.0
MarkupSafe==2.0
python-dateutil==2.8.1
pytz==2019.3
six==1.15.0
urllib3==1.25.8


